+++ NEWS.md
@@ -1,3 +1,12 @@
+# routr 0.4.0
+
+* Add `get_handler()` method to `Route` (#9, @cpsievert)
+* Add `root` field to `Route` which will get appended to all paths before 
+  matching to an incomming request
+* Add `remap_handlers()` to loop through all handlers and reassign them based on
+  a user provided function. (#8)
+* Added pkgdown site at https://routr.data-imaginist.com
+
 # routr 0.3.0
 
 * Add `on_error()` method to modify how errors are handled. The default is now